Add files via upload
This is version 0.8.15 of UAE, the Un*x Amiga Emulator.

Versions numbered 0.8.x are beta versions, to be used for development only.
If you are only a user of UAE, you are better off using the latest 0.7.x
version or waiting for 0.9.0.


License
=======

UAE is available under the terms of the GNU General Public License. This means
that it is free software: you are welcome to distribute copies of it and/or
modify it, under certain conditions. It also means that there is no warranty
for UAE.
See the file COPYING that is included in the top level directory of this
archive for details.

The SCSI code distributed with UAE is also available under the terms of the
GPL and was written by J. Schilling.